+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - examples of "Lattice Surgery"
 - 'measure' method of QState, Stabilizer (one shot measuremant in computational basis)
 ### Changed
-- QComp ('reset' method -> 'clear' method, redefine 'reset' method as adding non-unitary reset gate, return value of 'run' method)
+- specifications of QComp
 ### Fixed
 - 'measure' method of QComp for qlazy's qstate_simulator